fix: consolidate TMessage import from custom-messages in chat-service, chat-session, and message components
